Move forecasts related stuff into app/forecasts
